PokemonBag/requirements.txt
Flask
requests
pokebase
pypokedex
Pillow
playsound
